+++ software/rpi-direct/tests/motion/main.cpp
@@ -16,7 +16,7 @@ int main(){
     auto end_time = std::chrono::high_resolution_clock::now();
 
     for(float i = 0; i < 2*PI; i += (5.0*PI/180.0)){
-        std::tuple<int, int, int, int> motor_speeds = motion_controller.translate(i, 1);
+        std::tuple<float, float, float, float> motor_speeds = motion_controller.translate(i, 1);
         std::cout << i << std::endl;
         std::cout << std::get<0>(motor_speeds) << " " << std::get<1>(motor_speeds) << " " << std::get<2>(motor_speeds) << " " << std::get<3>(motor_speeds) << std::endl;
         std::cout << "--------------------------------" << std::endl;
